FM test tweaks

+++ ffs/tests/context_test2.c
@@ -19,11 +19,13 @@
 #include "ffs.h"
 
 #include "test_funcs.h"
+#include <sys/types.h>
+#include <sys/stat.h>
 
-static void test_receive(char *buffer, int buf_size, int finished,
+static void test_receive(char *buffer, size_t buf_size, int finished,
 			       int test_level);
-static void test_all_receive(char *buffer, int buf_size, int finished);
-static void write_buffer(char *buf, int size);
+static void test_all_receive(char *buffer, size_t buf_size, int finished);
+static void write_buffer(char *buf, size_t size);
 static void read_test_only();
 
 static int write_output = 0;
@@ -171,7 +173,7 @@ main(int argc, char **argv)
 #endif
 
 static char *
-get_buffer(int *size_p)
+get_buffer(size_t *size_p)
 {
     static int file_fd = 0;
     static char *buffer = NULL;
@@ -233,7 +235,7 @@ static void
 read_test_only()
 {
     char *input;
-    int size;
+    size_t size;
     while ((input = get_buffer(&size)) != NULL) {
 	test_all_receive(input, size, 0);
     }
@@ -257,26 +259,26 @@ set_targets(FFSContext context)
     reader_register_ioformat = FFSset_fixed_target(context, reader_register_format_list);
 }
 
-int base_size_func(FFSContext context, char *src, int rec_len,
-		   int native_struct_size)
+size_t base_size_func(FFSContext context, char *src, size_t rec_len,
+		   size_t native_struct_size)
 {
     return native_struct_size;
 }
 
-int total_size_func(FFSContext context, char *src, int rec_len, 
-		    int native_struct_size)
+size_t total_size_func(FFSContext context, char *src, size_t rec_len, 
+		    size_t native_struct_size)
 {
     return FFS_est_decode_length(context, src, rec_len);
 }
 
 static int 
-decode_in_place(FFSContext context, char *src, int src_len, void *dest)
+decode_in_place(FFSContext context, char *src, size_t src_len, void *dest)
 {
     if (FFSdecode_in_place_possible(FFSTypeHandle_from_encode(context, src))) {
 	int ret, header_len;
 	char *real_dest;
 	ret = FFSdecode_in_place(context, src, (void**)&real_dest);
-	header_len = real_dest - src;
+	header_len = (int)(intptr_t)(real_dest - src);
 	memcpy(dest, real_dest, src_len - header_len);
 	return ret;
     } else {
@@ -285,22 +287,22 @@ decode_in_place(FFSContext context, char *src, int src_len, void *dest)
 }
 
 static int
-decode_IOcontext_wrapper(FFSContext context, char *src, int src_len, void *dest)
+decode_IOcontext_wrapper(FFSContext context, char *src, size_t src_len, void *dest)
 {
     return FFSdecode(context, src, dest);
 }
 
 static int
-decode_to_buffer_IOcontext_wrapper(FFSContext context, char *src, int src_len,
+decode_to_buffer_IOcontext_wrapper(FFSContext context, char *src, size_t src_len,
 				   void *dest)
 {
     return FFSdecode_to_buffer(context, src, dest);
 }
 
-typedef int (*size_func_t)(FFSContext context, char *src, int buf_size, 
-				   int nativ_struct);
+typedef size_t (*size_func_t)(FFSContext context, char *src, size_t buf_size, 
+				   size_t nativ_struct);
 
-typedef int (*decode_func_t)(FFSContext context, char *src, int src_len, 
+typedef int (*decode_func_t)(FFSContext context, char *src, size_t src_len, 
 				   void *dest);
 
 size_func_t size_funcs[] = {base_size_func, total_size_func, total_size_func};
@@ -312,7 +314,7 @@ decode_func_t decode_funcs[] = {decode_IOcontext_wrapper,
 #define NUM_TESTS 3
 
 static void
-test_all_receive(char *buffer, int buf_size, int finished)
+test_all_receive(char *buffer, size_t buf_size, int finished)
 {
     int test_type = 0;
     char *tmp_buffer = malloc(buf_size);
@@ -324,7 +326,7 @@ test_all_receive(char *buffer, int buf_size, int finished)
 }
 	
 static void*
-get_mem(int size)
+get_mem(size_t size)
 {
     char *buffer;
     unsigned int beef = 0xdeadbeef;
@@ -335,7 +337,7 @@ get_mem(int size)
 }
 
 static void
-check_mem(int size, char *buffer)
+check_mem(size_t size, char *buffer)
 {
     unsigned int beef = 0xdeadbeef;
     if (memcmp(buffer+size, &beef, 4) != 0) {
@@ -345,7 +347,7 @@ check_mem(int size, char *buffer)
 
     
 static void
-test_receive(char *buffer, int buf_size, int finished, int test_level)
+test_receive(char *buffer, size_t buf_size, int finished, int test_level)
 {
     static FFSContext c = NULL;
     static int node_rec_count[NUM_TESTS] = {0,0,0};
@@ -368,7 +370,7 @@ test_receive(char *buffer, int buf_size, int finished, int test_level)
 	}
 	if (((test_only == NULL) || (strcmp(test_only, "node") == 0)) &&
 		   (buffer_format == node_ioformat)) {
-	    int size = size_func(rcv_context, buffer, buf_size, 
+	    size_t size = size_func(rcv_context, buffer, buf_size, 
 				 sizeof(add_rec));
 	    node_ptr read_data = get_mem(size);
 	    struct visit_table v;
@@ -390,7 +392,7 @@ test_receive(char *buffer, int buf_size, int finished, int test_level)
 	    node_rec_count[test_level]++;
 	} else 	if (((test_only == NULL) || (strcmp(test_only, "psa") == 0)) &&
 		   (buffer_format == psa_ioformat)) {
-	    int size = size_func(rcv_context, buffer, buf_size, 
+	    size_t size = size_func(rcv_context, buffer, buf_size, 
 				 sizeof(add_rec));
 	    pointer_to_static_array_ptr read_data = get_mem(size);
 	    memset(read_data, 0, size);
@@ -404,7 +406,7 @@ test_receive(char *buffer, int buf_size, int finished, int test_level)
 	    free(read_data);
 	} else 	if (((test_only == NULL) || (strcmp(test_only, "reader_register") == 0)) &&
 		   (buffer_format == reader_register_ioformat)) {
-	    int size = size_func(rcv_context, buffer, buf_size, 
+	    size_t size = size_func(rcv_context, buffer, buf_size, 
 				 sizeof(add_rec));
 	    struct _reader_register_msg *read_data = get_mem(size);
 	    memset(read_data, 0, size);
@@ -439,7 +441,7 @@ test_receive(char *buffer, int buf_size, int finished, int test_level)
 }
 
 static void
-write_buffer(char *buf, int size)
+write_buffer(char *buf, size_t size)
 {
     static int file_fd = 0;
     unsigned short ssize;
@@ -449,11 +451,11 @@ write_buffer(char *buf, int size)
     if (file_fd == 0) {
 	file_fd = open(output_file, O_WRONLY|O_CREAT|O_TRUNC|O_BINARY, 0777);
     }
-    printf("Writing buffer of size %d\n", size);
-    ssize = size;
+    printf("Writing buffer of size %zd\n", size);
+    ssize = (unsigned short) size;
     csize = ssize & 0xff;
     if (write(file_fd, &csize, 1) != 1) exit(1);/* low byte of 2-byte size */
     csize = ((ssize >> 8) & 0xff);
     if (write(file_fd, &csize, 1) != 1) exit(1);/* high byte of 2-byte size */
-    if (write(file_fd, buf, size) != size) exit(1);
+    if (write(file_fd, buf, (int)size) != size) exit(1);
 }
